Segregate component files
